tests/submission/W3C/domhtml/HTMLQuoteElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - QUOTE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-blockquote-element.html#dom-quote-cite" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<q cite="/support/html/sample.html">The Q element is intended for short quotations</q>
</p>
<blockquote cite="/support/html/sample.html">
<p>The BLOCKQUOTE element is used for long quotations.</p>
</blockquote>



<script>
var domtest = async_test("The cite IDL attribute must reflect the element's cite content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcite;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("blockquote");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcite = testNode.cite;

      assert_uri_equals(document.location.protocol,null,document.location.host,"sample.html",null,null,null,true,vcite);

    domtest.done();
}));

</script>
<p>The cite IDL attribute must reflect the element's cite content attribute.</p>
<div id="log"></div>
</body></html>
